Remove unused Discord button and message handlers
Deleted multiple files containing outdated button and message handler implementations for Discord bot functionality, including payment and detail handling logic. This cleanup removes redundant or unused code to improve maintainability and reduce complexity.
