+++ internal/discord/commands/.placeholder
@@ -1,0 +1,1 @@
+This directory contains command definitions.
