+++ internal/discord/handlers/.placeholder
@@ -1,0 +1,1 @@
+This directory contains handler implementations.
