+++ internal/discord/handlers/components.go
@@ -1,20 +1,22 @@
-package discord
+package handlers
 
 import (
-	"github.com/bwmarrin/discordgo"
 	"log"
 	"strings"
+
+	"github.com/bwmarrin/discordgo"
 )
 
+// Component Custom IDs - shared constants for all interactive components
 const (
-	// Component Custom IDs
 	payDebtButtonPrefix        = "pay_debt_"
 	viewDetailButtonPrefix     = "view_detail_"
 	requestPaymentButtonPrefix = "request_payment_"
 	markPaidButtonPrefix       = "mark_paid_"
 	confirmPaymentButtonPrefix = "confirm_payment_"
-	viewDuesButtonPrefix       = "view_dues_"
 	billAllocateButtonPrefix   = "bill_allocate_"
+	billSkipButtonPrefix       = "bill_skip_"
+	billCancelButtonID         = "bill_cancel"
 	billUsersSelectPrefix      = "bill_users_select_"
 	debtDropdownID             = "debt_dropdown"
 )
@@ -39,29 +41,25 @@ func handleMessageComponentInteraction(s *discordgo.Session, i *discordgo.Intera
 		handlePayDebtButton(s, i)
 	case strings.HasPrefix(customID, viewDetailButtonPrefix):
 		handleViewDetailButton(s, i)
+	case strings.HasPrefix(customID, requestPaymentButtonPrefix):
+		handleRequestPaymentButton(s, i)
 	case strings.HasPrefix(customID, markPaidButtonPrefix):
 		handleMarkPaidButton(s, i)
 	case strings.HasPrefix(customID, confirmPaymentButtonPrefix):
 		handleConfirmPaymentButton(s, i)
-	case strings.HasPrefix(customID, requestPaymentButtonPrefix):
-		handleRequestPaymentButton(s, i)
-	case strings.HasPrefix(customID, viewDuesButtonPrefix):
-		handleViewDuesButton(s, i)
-	case strings.HasPrefix(customID, billAllocateButtonPrefix):
-		handleBillAllocateButton(s, i)
-	case strings.HasPrefix(customID, billUsersSelectPrefix):
-		handleUserSelectSubmit(s, i)
 	case customID == debtDropdownID:
 		handleDebtDropdown(s, i)
+	case strings.HasPrefix(customID, billAllocateButtonPrefix):
+		handleBillAllocateButton(s, i)
+	case strings.HasPrefix(customID, billSkipButtonPrefix):
+		handleBillSkipButton(s, i)
+	case customID == billCancelButtonID:
+		handleBillCancelButton(s, i)
+	case strings.HasPrefix(customID, billUsersSelectPrefix):
+		handleUserSelectSubmit(s, i)
 	default:
 		log.Printf("Unknown component interaction: %s", customID)
-		s.InteractionRespond(i.Interaction, &discordgo.InteractionResponse{
-			Type: discordgo.InteractionResponseChannelMessageWithSource,
-			Data: &discordgo.InteractionResponseData{
-				Content: "ไม่รู้จัก interaction นี้ โปรดติดต่อผู้ดูแลระบบ",
-				Flags:   discordgo.MessageFlagsEphemeral,
-			},
-		})
+		respondWithError(s, i, "ไม่รู้จัก interaction นี้ โปรดติดต่อผู้ดูแลระบบ")
 	}
 }
 
@@ -84,12 +82,6 @@ func handleModalSubmit(s *discordgo.Session, i *discordgo.InteractionCreate) {
 		handlePayDebtModalSubmit(s, i)
 	} else {
 		log.Printf("Unknown modal interaction: %s", customID)
-		s.InteractionRespond(i.Interaction, &discordgo.InteractionResponse{
-			Type: discordgo.InteractionResponseChannelMessageWithSource,
-			Data: &discordgo.InteractionResponseData{
-				Content: "ไม่รู้จัก modal interaction นี้ โปรดติดต่อผู้ดูแลระบบ",
-				Flags:   discordgo.MessageFlagsEphemeral,
-			},
-		})
+		respondWithError(s, i, "ไม่รู้จัก modal interaction นี้ โปรดติดต่อผู้ดูแลระบบ")
 	}
 }
